Support uploading with presigned urls
Adds S3::PresignedUploader to upload files to Scholarsphere via
presigned urls. A request is sent to Scholarsphere for a url to upload
data that doesn't require additional AWS credentials. Once the url is
returned, the file can be uploaded to the bucket, which includes a
prefix subdirectory. After the file is successfully uploaded, additional
requests can be sent to ingest the file and its associated metadata
into the Scholarsphere repository.