Support uploading with presigned urls
Drops support for the current multipart uploader in favor of using
presigned urls exclusively.

A request is sent to Scholarsphere for a url to upload data that doesn't
require additional AWS credentials. Once the url is returned, the file
can be uploaded to the bucket, which includes a prefix subdirectory.
After the file is successfully uploaded, additional requests can be sent
to ingest the file and its associated metadata into the Scholarsphere
repository.

Note that this uploader does _not_ support multipart uploading. That
will be re-implemented at a later date using the same presigned url
approach.