Add .circleci/config.yml
In order to make CI a bit easier to manage, we're checking-in
Gemfile.lock. For gems, this is sometimes considered a bad
practice, however, we think this isn't the case. The lock file will only influence
local development, which is a good thing: Now developers will have the
same gems in their environments as each other. When the gem is installed
into a project, the gemspec will determine which gems are installed, and
not the lock file. Additionally, we'll be immune from gem compatibility
issues which cause a build to suddenly start to fail.

In order to be aware of compatibility issues, we can schedule a separate
build that runs without the lock file, so it uses all the latest
versions specified in the gemspec, but does not block current
development.