Enable md5 checksum verification with presigned urls
In order to for S3 to verify a file's checksum with uploading with
presigned url, the checksum must be sent as part of the presigned url
request. Along with a corresponding update to the Scholarsphere API, the
checksum is sent to the API when requesting the presigned url.

This also includes some changes to error reporting and the client will
now raise Error exceptions when encountering non-20x responses from
Scholarsphere or S3 in the requests leading up to the /ingest call. This
enables the caller to better diagnose errors along the way.